Placed utility code in its own folder
