+++ prototype/utils/bitmap.c
