+++ prototype/utils/bitmap.h
